Fixed up test script path issues
In order to make test scripts use the misc/plog.sh script for printing
their logs, I'd created a $rootdir variable in the bashscript so
these, and the makefile directory, could be easily referenced.

However, my method of getting the root directory was flawed if you
tried to get the directory from the same directory as the scripts.
This method should work more of the time.