Fixed up behaviour of hooks in Python
This was a fairly severe bug - Hooks treated ctypes._Pointer types as
integers or similar, which meant that they assumed that the C
hook_call function was called with a pointer to the pointer, just as
integers are called with a pointer to the integer. This would cause
Python to falsely dereference the pointer.

The only place where pointers are really used as hook parameters is
in the delete_struct hooks of Containers. This means that when a
StructObject is deleted, Python kept being passed invalid pointers
to the deleted object, which led to Python making calls on dead
objects and reading free'd memory.