Changed editor.bind to use Mode rather than core.keymap.
Since bindings no longer depend on core.keymap, it should be safe to finally
start dismantling it, and replacing it with the C-modes system.

The next stage is to replace per-buffer keyframes with some kind of
mode-manager for per-buffer modes, which should be interesting. I'm not yet
decided on what language to put that in. It certainly shouldn't need speed, so
the main consideration should be whether it needs to be availible to anything
in languages other than Python.