Removed KeyFrames etc. from core.keymap. Now uses Container.
Also modified HookFunct so that it takes its hook as an argument rather than
trying to inherit it from its parent. This allows it to be used as a base class
for other similar things.

 * KeymapFunct now inherits from HookFunct and overrides only its call method.
 * Added KeymapObj(StructObject) and KeymapFreeOnDel(KeymapObj) as struct
   representors of keymaps.
 * Added KeymapContainer(Container) to contain instances of Keymaps.
 * Keymap is now an instance of KeymapContainer
 * Removed everything else.
This commit has almost certainly caused a log of damage and broken a lot of
things, so prepare for a heavy damage control, bugfixing and rebuilding session.