First draft of ui/text.c
This file is intended to replace a decent chunk of ui/util, and generalize the
handling of text strings (with faces) in a nice way. It's also the first code in
edil to actually attempt unicode support, and hopefully should be the hardest
part.

Next up is to re-write the text drawing functions, support wide ncurses, which
will probably be a simple change of linked library, and finally to have python
encode its bytestrings in utf-8 instead of ascii.

This branch will probably finish changes to ui/text.c, and the text drawing
functions, stabilize and merge into master, then have a utf-8 branch fork from
master where we can do fun actual things with unicode.