Added doc/c-guidelines
The way that multi-line function arguments are displayed is different in the
guidelines. At the moment, the code does not comply with the guidelines.
mode.c was changed to comply, since it is unique to master at this time.

After pulling in various feature branches, I should go on a documenting and
stylizing spree and attack all my code with these guidelines.